requirements.txt
pyodbc==4.0.30
azure-identity==1.3.1
azure-keyvault-secrets==4.1.0